Rename api files to services and move them to services folder
