+++ UniVerseUI/universe-ui/src/services/eventsService.js
@@ -1,4 +1,4 @@
-import axios from "./axios"
+import axios from "../api/axios"
 
 export const getEvents = async () =>{
     const response = await axios.get('/events');
